+++ lib/core/rest.ts
@@ -1,7 +1,9 @@
 import { Application, Context, Router, Status } from "@oak/oak"
 import { PupApi } from "./api.ts"
 import { Pup } from "./pup.ts"
-import { generateKey, verifyJWT } from "@cross/jwt"
+import { generateKey } from "@cross/jwt"
+import { DEFAULT_REST_API_HOSTNAME, DEFAULT_REST_API_PORT, DEFAULT_SECRET_KEY_ALGORITHM } from "./configuration.ts"
+import { ValidateToken } from "../common/token.ts"
 
 const generateAuthMiddleware = (key: CryptoKey) => {
   return async (ctx: Context, next: () => Promise<unknown>) => {
@@ -20,9 +22,13 @@ const generateAuthMiddleware = (key: CryptoKey) => {
     }
     const token = parts[1]
     try {
-      const payload = await verifyJWT(token, key)
-      ctx.state.user = payload.user // Add user info to context state
-      await next() // Proceed if valid
+      const payload = await ValidateToken(token, key)
+      if (payload) {
+        await next() // Proceed if valid
+      } else {
+        ctx.response.status = Status.Unauthorized
+        ctx.response.body = { message: "Invalid/expired token" }
+      }
     } catch (_err) {
       ctx.response.status = Status.Unauthorized
       ctx.response.body = { message: "Invalid token" }
@@ -36,35 +42,58 @@ export class RestApi {
   private router: Router
   private appAbortController: AbortController
 
-  private port: number
-  private hostname: string
   private secret: string
   private key?: CryptoKey
+
+  public port: number
+  public hostname: string
 
   constructor(pup: Pup, port: number | undefined, hostname: string | undefined, jwtSecret: string) { // Takes a Pup instance
     this.pupApi = new PupApi(pup)
     this.app = new Application()
     this.router = new Router()
     this.appAbortController = new AbortController()
-    this.port = port || 16421
-    this.hostname = hostname || "localhost"
+    this.port = port || DEFAULT_REST_API_PORT
+    this.hostname = hostname || DEFAULT_REST_API_HOSTNAME
     this.secret = jwtSecret
     this.setupRoutes() // Setup routes within the constructor
   }
 
   private async setupKey() {
-    return await generateKey(this.secret, "HS512")
+    return await generateKey(this.secret, DEFAULT_SECRET_KEY_ALGORITHM)
   }
 
-  private async setupRoutes() {
+  private setupRoutes() {
+    this.router.get("/wss", (ctx) => {
+      if (!ctx.isUpgradable) {
+        ctx.throw(501)
+      }
+      const ws = ctx.upgrade()
+      if (!ctx.isUpgradable) {
+        ctx.throw(501)
+      }
+      const proxyFn = (d: unknown) => {
+        ws.send(JSON.stringify({ t: "log", d: d }))
+      }
+      this.pupApi.events.on("log", proxyFn)
+      /*ws.onopen = () => {
+        ws.send("Hello from server!")
+      }*/
+      ws.onmessage = (m) => {
+        ws.send(m.data as string)
+      }
+      ws.onclose = (_e) => {
+        this.pupApi.events.off("log", proxyFn)
+      }
+    })
+
     // Process related routes
     this.router
       .get("/processes", (ctx) => {
         ctx.response.body = this.pupApi.allProcessStates()
       })
       .get("/state", (ctx) => {
-        const ProcessStatees = this.pupApi.applicationState()
-        ctx.response.body = ProcessStatees
+        ctx.response.body = this.pupApi.applicationState()
       })
       .post("/processes/:id/start", (ctx) => {
         const id = ctx.params.id
@@ -116,93 +145,100 @@ export class RestApi {
           ctx.response.body = { error: err.message }
         }
       })
-
-    // Application State route
-    this.router.get("/application-state", (ctx) => {
-      ctx.response.body = this.pupApi.applicationState()
-    })
-
-    // Termination route
-    this.router.post("/terminate", (ctx) => {
-      // Add logic to read forceQuitMs from the request body if needed
-      const forceQuitMs = 3000 // Example value
-      try {
-        ctx.response.status = Status.OK
-      } catch (err) {
-        ctx.response.status = Status.InternalServerError
-        ctx.response.body = { error: err.message }
-      }
-      this.pupApi.terminate(forceQuitMs)
-    })
-
-    // Logging routes
-    this.router
-    /*.post("/log", async (ctx) => {
+      .post("/telemetry", async (ctx) => {
+        if (ctx.request.hasBody) {
+          try {
+            const parsedBody = await ctx.request.body.json()
+            const success = this.pupApi.telemetry(parsedBody)
+            if (success) {
+              ctx.response.status = Status.OK
+            } else {
+              ctx.response.status = Status.InternalServerError
+              ctx.response.body = { error: "Invalid data" }
+            }
+          } catch (err) {
+            ctx.response.status = Status.InternalServerError
+            ctx.response.body = { error: err.message }
+          }
+        }
+      })
+      .post("/terminate", (ctx) => {
+        // Add logic to read forceQuitMs from the request body if needed
+        const forceQuitMs = 3000 // Example value
+        try {
+          ctx.response.status = Status.OK
+        } catch (err) {
+          ctx.response.status = Status.InternalServerError
+          ctx.response.body = { error: err.message }
+        }
+        this.pupApi.terminate(forceQuitMs)
+      })
+      /*.post("/log", async (ctx) => {
         // Read severity, plugin, and message from request body
         const body = await ctx.request.body().value
         this.pupApi.log(body.severity, body.plugin, body.message)
         ctx.response.status = Status.Created
       })*/
-    this.router.get("/logs", async (context) => {
-      try {
-        const params = context.request.url.searchParams
+      .get("/logs", async (context) => {
+        try {
+          const params = context.request.url.searchParams
 
-        const processId = params.get("processId")
-        const startTimeStampParam = params.get("startTimeStamp")
-        const endTimeStampParam = params.get("endTimeStamp")
-        const severity = params.get("severity")
-        const nRowsParam = params.get("nRows")
+          const processId = params.get("processId")
+          const startTimeStampParam = params.get("startTimeStamp")
+          const endTimeStampParam = params.get("endTimeStamp")
+          const severity = params.get("severity")
+          const nRowsParam = params.get("nRows")
 
-        // Convert nRows to integer and validate
-        let nRows
-        if (nRowsParam) {
-          nRows = parseInt(nRowsParam, 10)
-          if (isNaN(nRows)) {
-            context.response.status = 400
-            context.response.body = { error: "nRows should be a number" }
-            return
+          // Convert nRows to integer and validate
+          let nRows
+          if (nRowsParam) {
+            nRows = parseInt(nRowsParam, 10)
+            if (isNaN(nRows)) {
+              context.response.status = 400
+              context.response.body = { error: "nRows should be a number" }
+              return
+            }
           }
-        }
-        let startTimeStamp
-        if (startTimeStampParam) {
-          startTimeStamp = parseInt(startTimeStampParam, 10)
-          if (isNaN(startTimeStamp)) {
-            context.response.status = 400
-            context.response.body = { error: "startTimeStamp should be a number" }
-            return
+          let startTimeStamp
+          if (startTimeStampParam) {
+            startTimeStamp = parseInt(startTimeStampParam, 10)
+            if (isNaN(startTimeStamp)) {
+              context.response.status = 400
+              context.response.body = { error: "startTimeStamp should be a number" }
+              return
+            }
           }
-        }
-        let endTimeStamp
-        if (endTimeStampParam) {
-          endTimeStamp = parseInt(endTimeStampParam, 10)
-          if (isNaN(endTimeStamp)) {
-            context.response.status = 400
-            context.response.body = { error: "endTimeStamp should be a number" }
-            return
+          let endTimeStamp
+          if (endTimeStampParam) {
+            endTimeStamp = parseInt(endTimeStampParam, 10)
+            if (isNaN(endTimeStamp)) {
+              context.response.status = 400
+              context.response.body = { error: "endTimeStamp should be a number" }
+              return
+            }
           }
+          const nRowsCapped = (!nRows || (nRows && nRows > 100)) ? 100 : nRows
+
+          let logContents = await this.pupApi.getLogs(processId || undefined, startTimeStamp, endTimeStamp, nRowsCapped)
+
+          if (severity) {
+            const severityLower = severity.toLowerCase()
+            logContents = logContents.filter((log) => log.severity.toLowerCase() === severityLower)
+          }
+
+          context.response.body = logContents
+        } catch (error) {
+          context.response.status = 500
+          context.response.body = { error: "Internal Server Error", message: error.message }
         }
-        const nRowsCapped = (!nRows || (nRows && nRows > 100)) ? 100 : nRows
-
-        let logContents = await this.pupApi.getLogs(processId || undefined, startTimeStamp, endTimeStamp, nRowsCapped)
-
-        if (severity) {
-          const severityLower = severity.toLowerCase()
-          logContents = logContents.filter((log) => log.severity.toLowerCase() === severityLower)
-        }
-
-        context.response.body = logContents
-      } catch (error) {
-        context.response.status = 500
-        context.response.body = { error: "Internal Server Error", message: error.message }
-      }
-    })
-
-    this.app.use(generateAuthMiddleware(await this.setupKey()))
-    this.app.use(this.router.routes())
-    this.app.use(this.router.allowedMethods())
+      })
   }
 
   public async start() {
+    this.app.use(generateAuthMiddleware(await this.setupKey()))
+    this.app.use(this.router.routes())
+    this.app.use(this.router.allowedMethods())
+
     this.pupApi.log("info", "rest", `Starting the REST API`)
     //await this.app.listen({ port, signal: this.appAbortController.signal })
     await this.app.listen({ port: this.port, hostname: this.hostname, signal: this.appAbortController.signal })