+++ plugins/web-interface/deps.ts
@@ -1,3 +1,0 @@
-export { Application, Router } from "@oak/oak"
-export { dirname, fromFileUrl } from "@std/path"
-export { Bundlee } from "@hexagon/bundlee"